# protocol/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pulse upload protocol",
    "description": "The HTTP surface of the Pulse upload protocol, as implemented by @mieweb/pulsevault. Generated from the plugin's route schemas — see PROTOCOL.md for the full contract.",
    "version": "2.3"
  },
...
  "paths": {
    "/pulsevault/upload": {
      "post": {
        "summary": "TUS resumable upload endpoint",
        "tags": [
          "pulsevault"
        ],
        "description": "TUS v1 resumable upload protocol.\n\n- `POST` creates a new upload. The `Upload-Metadata` header must include base64-encoded key/value pairs:\n  - `artifactId` (or the legacy `videoid`/`projectid` aliases) — a UUID generated by your server.\n  - `filename` — original filename; the extension must match the kind's allowed list.\n  - `kind` — `video` (default), `project`, `captions`, or `thumbnail`. Determines the storage subdir and which completion hooks fire.\n  - `relatedTo` — optional UUID of another artifact this one belongs to (e.g. the captions, beat manifest or thumbnail belonging to a pulse's video).\n  - `checksum` — optional `<algorithm>:<hex digest>` of the finished file, verified post-upload if a checksum validator is configured.\n- `PATCH` appends a chunk at the offset given by `Upload-Offset`, with `Content-Type: application/offset+octet-stream`.\n- `HEAD` returns the current offset for a resumable upload.\n- `DELETE` removes the upload and its artifact, whether in flight (a cancel) or finished. Authorized as `delete`.\n\nSee https://tus.io/protocols/resumable-upload for the full protocol.",
        "responses": {
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Authorize hook rejected the request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Authorize hook rejected the request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "TUS resumable upload endpoint",
        "tags": [
          "pulsevault"
        ],
        "description": "TUS v1 resumable upload protocol.\n\n- `POST` creates a new upload. The `Upload-Metadata` header must include base64-encoded key/value pairs:\n  - `artifactId` (or the legacy `videoid`/`projectid` aliases) — a UUID generated by your server.\n  - `filename` — original filename; the extension must match the kind's allowed list.\n  - `kind` — `video` (default), `project`, `captions`, or `thumbnail`. Determines the storage subdir and which completion hooks fire.\n  - `relatedTo` — optional UUID of another artifact this one belongs to (e.g. the captions, beat manifest or thumbnail belonging to a pulse's video).\n  - `checksum` — optional `<algorithm>:<hex digest>` of the finished file, verified post-upload if a checksum validator is configured.\n- `PATCH` appends a chunk at the offset given by `Upload-Offset`, with `Content-Type: application/offset+octet-stream`.\n- `HEAD` returns the current offset for a resumable upload.\n- `DELETE` removes the upload and its artifact, whether in flight (a cancel) or finished. Authorized as `delete`.\n\nSee https://tus.io/protocols/resumable-upload for the full protocol.",
        "responses": {
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Authorize hook rejected the request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Authorize hook rejected the request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "head": {
        "summary": "TUS resumable upload endpoint",
        "tags": [
          "pulsevault"
        ],
        "description": "TUS v1 resumable upload protocol.\n\n- `POST` creates a new upload. The `Upload-Metadata` header must include base64-encoded key/value pairs:\n  - `artifactId` (or the legacy `videoid`/`projectid` aliases) — a UUID generated by your server.\n  - `filename` — original filename; the extension must match the kind's allowed list.\n  - `kind` — `video` (default), `project`, `captions`, or `thumbnail`. Determines the storage subdir and which completion hooks fire.\n  - `relatedTo` — optional UUID of another artifact this one belongs to (e.g. the captions, beat manifest or thumbnail belonging to a pulse's video).\n  - `checksum` — optional `<algorithm>:<hex digest>` of the finished file, verified post-upload if a checksum validator is configured.\n- `PATCH` appends a chunk at the offset given by `Upload-Offset`, with `Content-Type: application/offset+octet-stream`.\n- `HEAD` returns the current offset for a resumable upload.\n- `DELETE` removes the upload and its artifact, whether in flight (a cancel) or finished. Authorized as `delete`.\n\nSee https://tus.io/protocols/resumable-upload for the full protocol.",
        "responses": {
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Authorize hook rejected the request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Authorize hook rejected the request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "TUS resumable upload endpoint",
        "tags": [
          "pulsevault"
        ],
        "description": "TUS v1 resumable upload protocol.\n\n- `POST` creates a new upload. The `Upload-Metadata` header must include base64-encoded key/value pairs:\n  - `artifactId` (or the legacy `videoid`/`projectid` aliases) — a UUID generated by your server.\n  - `filename` — original filename; the extension must match the kind's allowed list.\n  - `kind` — `video` (default), `project`, `captions`, or `thumbnail`. Determines the storage subdir and which completion hooks fire.\n  - `relatedTo` — optional UUID of another artifact this one belongs to (e.g. the captions, beat manifest or thumbnail belonging to a pulse's video).\n  - `checksum` — optional `<algorithm>:<hex digest>` of the finished file, verified post-upload if a checksum validator is configured.\n- `PATCH` appends a chunk at the offset given by `Upload-Offset`, with `Content-Type: application/offset+octet-stream`.\n- `HEAD` returns the current offset for a resumable upload.\n- `DELETE` removes the upload and its artifact, whether in flight (a cancel) or finished. Authorized as `delete`.\n\nSee https://tus.io/protocols/resumable-upload for the full protocol.",
        "responses": {
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Authorize hook rejected the request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Authorize hook rejected the request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "options": {
        "summary": "TUS resumable upload endpoint",
        "tags": [
          "pulsevault"
        ],
        "description": "TUS v1 resumable upload protocol.\n\n- `POST` creates a new upload. The `Upload-Metadata` header must include base64-encoded key/value pairs:\n  - `artifactId` (or the legacy `videoid`/`projectid` aliases) — a UUID generated by your server.\n  - `filename` — original filename; the extension must match the kind's allowed list.\n  - `kind` — `video` (default), `project`, `captions`, or `thumbnail`. Determines the storage subdir and which completion hooks fire.\n  - `relatedTo` — optional UUID of another artifact this one belongs to (e.g. the captions, beat manifest or thumbnail belonging to a pulse's video).\n  - `checksum` — optional `<algorithm>:<hex digest>` of the finished file, verified post-upload if a checksum validator is configured.\n- `PATCH` appends a chunk at the offset given by `Upload-Offset`, with `Content-Type: application/offset+octet-stream`.\n- `HEAD` returns the current offset for a resumable upload.\n- `DELETE` removes the upload and its artifact, whether in flight (a cancel) or finished. Authorized as `delete`.\n\nSee https://tus.io/protocols/resumable-upload for the full protocol.",
        "responses": {
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Authorize hook rejected the request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Authorize hook rejected the request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/pulsevault/artifacts/{artifactId}/poster": {
      "get": {
        "summary": "Serve a video's poster frame",
        "tags": [
          "pulsevault"
        ],
        "description": "Streams or redirects to the finished thumbnail `relatedTo` the video, exactly as `GET /artifacts/:artifactId` would serve it. Runs the `authorize` hook with `phase: \"resolve\"` on the video, so whoever may watch it may see its poster. 404 until the poster has landed.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "token",
            "required": false,
            "description": "Optional bearer token for pre-authenticated watch links, forwarded to the `authorize` hook as `ctx.token`."
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "artifactId",
            "required": true,
            "description": "UUID of the video (the pulse's anchor artifact)."
          }
        ],
        "responses": {
          "400": {
            "description": "`artifactId` is not a valid UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "`artifactId` is not a valid UUID.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Authorize hook rejected the request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Authorize hook rejected the request.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "The video has no finished poster frame.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The video has no finished poster frame.",
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
...
}