# src/routes/pulsevault.ts
import { createRequire } from 'node:module';
import type {
  FastifyPluginAsync,
  FastifyPluginOptions,
  FastifyReply,
  FastifyRequest,
  FastifySchema,
} from 'fastify';
import type { PulseVaultCacheOptions } from '../app.js';
import { createPulseVaultCore, PROTOCOL_VERSION, type PulseVaultCore } from '../core.js';
import type {
  PulseVaultOnUploadComplete,
  PulseVaultOnArtifactEvent,
} from '../lib/pulsevaultTus.js';
import type { PulseVaultReplayOptions, PulseVaultWebReadyOptions } from '../lib/completion.js';
import type { PulseVaultValidatePayload } from '../lib/magic.js';
import { type PulseVaultAuthorize } from '../lib/authorize.js';
import type { PulseVaultIssueViewLink } from '../lib/view-links.js';
import type { PulseVaultAllowedExtensions } from '../lib/options.js';
import type { PulseVaultStorage, UploadKind } from '../storage/types.js';

export type {
  PulseVaultAuthorize,
  PulseVaultAuthorizeContext,
  PulseVaultAuthorizePhase,
} from '../lib/authorize.js';

const require = createRequire(import.meta.url);

/**
 * A schema from `protocol/schemas/` — the protocol's single source (PROTOCOL.md §7) — for use
 * as a route schema, without the JSON Schema `$schema`/`$id` keys Fastify's compilers don't take.
 */
function protocolSchema(name: string): Record<string, unknown> {
  const { $schema: _schema, $id: _id, ...schema } = require(
    `../../protocol/schemas/${name}.schema.json`,
  ) as Record<string, unknown>;
  return schema;
}

// Mirrored in the opt-in `./augment.ts` re-export (for consumers who `import
// "@mieweb/pulsevault/augment"`) and kept here too so this plugin typechecks
// standalone even if the consumer never imports that module.
declare module 'fastify' {
  interface FastifyRequest {
    pulseVault?: { artifactId: string; kind: UploadKind; relatedTo?: string };
  }
}

export type PulseVaultRoutesOptions = {
  storage: PulseVaultStorage;
  maxUploadSize: number;
  allowedExtensions: PulseVaultAllowedExtensions;
  cache?: PulseVaultCacheOptions;
  authorize?: PulseVaultAuthorize;
  validatePayload?: PulseVaultValidatePayload;
  onUploadComplete?: PulseVaultOnUploadComplete;
  onArtifactEvent?: PulseVaultOnArtifactEvent;
  issueViewLink?: PulseVaultIssueViewLink;
  pulseShape?: boolean;
  reclaim?: { idleSeconds?: number } | false;
  lockWhenReady?: boolean;
  webReady?: PulseVaultWebReadyOptions | boolean;
  replayCompletions?: PulseVaultReplayOptions | false;
  /** Receives the core this plugin built, so the (encapsulating) host plugin can expose it. */
  onCore?: (core: PulseVaultCore) => void;
} & FastifyPluginOptions;

// Adds the OpenAPI-flavored fields without pulling `@fastify/swagger` into this
// plugin's type surface; consumers who do register it get the real module
// augmentation and these become native.
type OpenApiRouteSchema = FastifySchema & {
  tags?: string[];
  summary?: string;
  description?: string;
};

const pulseVaultErrorResponse = {
  type: 'object',
  properties: {
    ok: { type: 'boolean' },
    error: { type: 'string' },
  },
  required: ['ok', 'error'],
};

const tusRouteSchema: OpenApiRouteSchema = {
  tags: ['pulsevault'],
  summary: 'TUS resumable upload endpoint',
  description:
    'TUS v1 resumable upload protocol.\n\n' +
    '- `POST` creates a new upload. The `Upload-Metadata` header must include base64-encoded key/value pairs:\n' +
    '  - `artifactId` (or the legacy `videoid`/`projectid` aliases) — a UUID generated by your server.\n' +
    "  - `filename` — original filename; the extension must match the kind's allowed list.\n" +
    '  - `kind` — `video` (default), `project`, `captions`, or `thumbnail`. Determines the storage subdir and which completion hooks fire.\n' +
    "  - `relatedTo` — optional UUID of another artifact this one belongs to (e.g. the captions, beat manifest or thumbnail belonging to a pulse's video).\n" +
    '  - `checksum` — optional `<algorithm>:<hex digest>` of the finished file, verified post-upload if a checksum validator is configured.\n' +
    '- `PATCH` appends a chunk at the offset given by `Upload-Offset`, with `Content-Type: application/offset+octet-stream`.\n' +
    '- `HEAD` returns the current offset for a resumable upload.\n' +
    '- `DELETE` removes the upload and its artifact, whether in flight (a cancel) or finished. Authorized as `delete`.\n\n' +
    'See https://tus.io/protocols/resumable-upload for the full protocol.',
  response: {
    400: { description: 'Invalid request.', ...pulseVaultErrorResponse },
    403: {
      description: 'Authorize hook rejected the request.',
      ...pulseVaultErrorResponse,
    },
  },
};

const artifactDeleteSchema: OpenApiRouteSchema = {
  tags: ['pulsevault'],
  summary: 'Delete an uploaded artifact',
  description:
    'Deletes all storage for an artifactId (bytes + sidecar metadata), regardless of kind. Runs the `authorize` hook with `phase: "delete"` before the adapter\'s `remove` is called. Returns 204 on success, 404 if the artifactId was unknown, 501 if the adapter does not implement `remove`.',
  params: {
    type: 'object',
    properties: {
      artifactId: {
        type: 'string',
        format: 'uuid',
        description: 'UUID of the upload to delete.',
      },
    },
    required: ['artifactId'],
  },
  response: {
    400: {
      description: '`artifactId` is not a valid UUID.',
      ...pulseVaultErrorResponse,
    },
    403: {
      description: 'Authorize hook rejected the request.',
      ...pulseVaultErrorResponse,
    },
    404: { description: 'Artifact not found.', ...pulseVaultErrorResponse },
    501: {
      description: 'Storage adapter does not implement delete.',
      ...pulseVaultErrorResponse,
    },
  },
};

const artifactGetSchema: OpenApiRouteSchema = {
  tags: ['pulsevault'],
  summary: 'Serve a previously uploaded artifact',
  description:
    "Resolves the `artifactId` through the configured storage adapter and either streams the bytes or redirects (for CDN-backed adapters). The artifact's kind (video, project, or captions) is resolved from storage, not the URL. Runs the `authorize` hook before resolve.",
  params: {
    type: 'object',
    properties: {
      artifactId: {
        type: 'string',
        format: 'uuid',
        description: 'UUID returned from the upload flow.',
      },
    },
    required: ['artifactId'],
  },
  querystring: {
    type: 'object',
    properties: {
      token: {
        type: 'string',
        description:
          'Optional bearer token for pre-authenticated watch links. Forwarded to the `authorize` hook as `ctx.token` so parent servers can validate it without a separate login step.',
      },
    },
  },
  response: {
    400: {
      description: '`artifactId` is not a valid UUID.',
      ...pulseVaultErrorResponse,
    },
    403: {
      description: 'Authorize hook rejected the request.',
      ...pulseVaultErrorResponse,
    },
    404: { description: 'Artifact not found.', ...pulseVaultErrorResponse },
  },
};

const viewLinkSchema: OpenApiRouteSchema = {
  tags: ['pulsevault'],
  summary: 'Mint a read-only view link for a finished artifact',
  description:
    'Returns a token that opens the artifact (and the artifacts `relatedTo` it) as `GET /artifacts/:artifactId?token=`, and nothing more — no uploads, deletes or further links. Runs the `authorize` hook with `phase: "share"`; the host\'s `issueViewLink` decides how long the link works. 404 unless the server enables view links (`/capabilities` `viewLinks`) and the artifact is finished; 403 when the host refuses a link for it.',
  params: {
    type: 'object',
    properties: {
      artifactId: {
        type: 'string',
        format: 'uuid',
        description: 'UUID of the finished artifact to link to.',
      },
    },
    required: ['artifactId'],
  },
  response: {
    200: protocolSchema('view-link'),
    400: {
      description: '`artifactId` is not a valid UUID.',
      ...pulseVaultErrorResponse,
    },
    403: {
      description: 'Authorize hook rejected the request, or the host refused a link for this artifact.',
      ...pulseVaultErrorResponse,
    },
    404: {
      description: 'View links are not enabled, or the artifact is not found or not finished.',
      ...pulseVaultErrorResponse,
    },
  },
};

const capabilitiesSchema: OpenApiRouteSchema = {
  tags: ['pulsevault'],
  summary: "Discover this deployment's protocol version and configuration",
  description:
    'Unauthenticated — the response carries no secrets. Lets a client detect protocol compatibility, allowed artifact kinds/extensions and the upload size cap before pairing.',
  response: { 200: protocolSchema('capabilities') },
};

// Fastify doesn't type route params from the schema alone, so `:artifactId` comes through as `unknown`.
const artifactStatusSchema: OpenApiRouteSchema = {
  tags: ['pulsevault'],
  summary: "Where an upload is: uploading, processing or ready",
  description:
    "Reports the artifact's state (`unknown`, `uploading`, `processing` while a web-ready conversion rewrites it, `ready`), its kind and `relatedTo`, the bytes received against its declared length, whether the host's `onUploadComplete` has finished (`acknowledged`), and what the host recorded with `recordOutcome` (`outcome`). Runs the `authorize` hook with `phase: \"status\"`; `createCapabilityAuthorize` grants it to the pairing token and to a view token. Never cached.",
  params: {
    type: 'object',
    properties: {
      artifactId: {
        type: 'string',
        format: 'uuid',
        description: 'UUID of the upload to report on.',
      },
    },
    required: ['artifactId'],
  },
  querystring: {
    type: 'object',
    properties: {
      token: {
        type: 'string',
        description:
          'Optional bearer token (the pairing token or a view token), forwarded to the `authorize` hook as `ctx.token` for browsers that can’t set a header.',
      },
    },
  },
  response: {
    200: protocolSchema('artifact-status'),
    400: {
      description: '`artifactId` is not a valid UUID.',
      ...pulseVaultErrorResponse,
    },
    403: {
      description: 'Authorize hook rejected the request.',
      ...pulseVaultErrorResponse,
    },
  },
};

const artifactPosterSchema: OpenApiRouteSchema = {
  tags: ['pulsevault'],
  summary: "Serve a video's poster frame",
  description:
    "Streams or redirects to the finished thumbnail `relatedTo` the video, exactly as `GET /artifacts/:artifactId` would serve it. Runs the `authorize` hook with `phase: \"resolve\"` on the video, so whoever may watch it may see its poster. 404 until the poster has landed.",
  params: {
    type: 'object',
    properties: {
      artifactId: {
        type: 'string',
        format: 'uuid',
        description: "UUID of the video (the pulse's anchor artifact).",
      },
    },
    required: ['artifactId'],
  },
  querystring: {
    type: 'object',
    properties: {
      token: {
        type: 'string',
        description:
          'Optional bearer token for pre-authenticated watch links, forwarded to the `authorize` hook as `ctx.token`.',
      },
    },
  },
  response: {
    400: {
      description: '`artifactId` is not a valid UUID.',
      ...pulseVaultErrorResponse,
    },
    403: {
      description: 'Authorize hook rejected the request.',
      ...pulseVaultErrorResponse,
    },
    404: { description: 'The video has no finished poster frame.', ...pulseVaultErrorResponse },
  },
};

type ArtifactIdParams = { artifactId?: unknown };

/** Coerce the raw route param to a string; an invalid UUID is rejected by the core's own check either way. */
function paramToString(value: unknown): string {
  return typeof value === 'string' ? value : '';
}

const pulseVaultRoutes: FastifyPluginAsync<PulseVaultRoutesOptions> = async (fastify, opts) => {
  const {
    storage,
    maxUploadSize,
    allowedExtensions,
    cache,
    authorize,
    validatePayload,
    onUploadComplete,
    onArtifactEvent,
    issueViewLink,
    pulseShape,
    reclaim,
    lockWhenReady,
    webReady,
    replayCompletions,
    onCore,
  } = opts;

  // All hook orchestration, tus glue, and artifact GET/DELETE logic lives in
  // the framework-agnostic core (`../core.js`); this file only wires Fastify's
  // routing/schema/lifecycle around it, so behavior can't drift from the
  // `./core` entry point non-Fastify hosts use.
  const core = createPulseVaultCore({
    storage,
    // `fastify.prefix` is `""` when the plugin is mounted at the root.
    basePath: fastify.prefix,
    maxUploadSize,
    allowedExtensions,
    cache,
    authorize,
    validatePayload,
    onUploadComplete,
    onArtifactEvent,
    issueViewLink,
    pulseShape,
    reclaim,
    lockWhenReady,
    webReady,
    replayCompletions,
    logger: fastify.log,
  });
  onCore?.(core);

  fastify.addContentTypeParser('application/offset+octet-stream', (_request, _payload, done) => {
    done(null);
  });

  const tusMethods = ['POST', 'PATCH', 'HEAD', 'DELETE', 'OPTIONS'] as const;
  const tusHandler = async (request: FastifyRequest, reply: FastifyReply) => {
    // The core writes directly to the raw response (tus dispatch, authorize
    // rejections, error translation) — hijack so Fastify doesn't also try to
    // manage this reply afterward.
    reply.hijack();
    await core.handleTus(request.raw, reply.raw);
  };
  fastify.route({
    method: [...tusMethods],
    url: '/upload',
    schema: tusRouteSchema,
    handler: tusHandler,
  });
  fastify.route({
    method: [...tusMethods],
    url: '/upload/*',
    schema: tusRouteSchema,
    handler: tusHandler,
  });

  fastify.get('/capabilities', { schema: capabilitiesSchema }, async (request, reply) => {
    reply.hijack();
    await core.handleCapabilities(request.raw, reply.raw);
  });

  fastify.delete(
    '/artifacts/:artifactId',
    { schema: artifactDeleteSchema },
    async (request, reply) => {
      const artifactId = paramToString((request.params as ArtifactIdParams)?.artifactId);
      reply.hijack();
      await core.handleArtifactDelete(request.raw, reply.raw, artifactId);
    },
  );

  fastify.post(
    '/artifacts/:artifactId/view-link',
    { schema: viewLinkSchema },
    async (request, reply) => {
      const artifactId = paramToString((request.params as ArtifactIdParams)?.artifactId);
      reply.hijack();
      await core.handleViewLink(request.raw, reply.raw, artifactId);
    },
  );

  fastify.get('/artifacts/:artifactId', { schema: artifactGetSchema }, async (request, reply) => {
    const artifactId = paramToString((request.params as ArtifactIdParams)?.artifactId);
    const token = (request.query as { token?: string })?.token;
    reply.hijack();
    await core.handleArtifactGet(request.raw, reply.raw, artifactId, token);
  });

  fastify.get(
    '/artifacts/:artifactId/status',
    { schema: artifactStatusSchema },
    async (request, reply) => {
      const artifactId = paramToString((request.params as ArtifactIdParams)?.artifactId);
      const token = (request.query as { token?: string })?.token;
      reply.hijack();
      await core.handleStatus(request.raw, reply.raw, artifactId, token);
    },
  );

  fastify.get(
    '/artifacts/:artifactId/poster',
    { schema: artifactPosterSchema },
    async (request, reply) => {
      const artifactId = paramToString((request.params as ArtifactIdParams)?.artifactId);
      const token = (request.query as { token?: string })?.token;
      reply.hijack();
      await core.handlePoster(request.raw, reply.raw, artifactId, token);
    },
  );
};

export type { PulseVaultCore };

export default pulseVaultRoutes;
export { PROTOCOL_VERSION };